% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENCODE}
\alias{ENCODE}
\title{Transcription factor targets estimated from ENCODE data.}
\format{A list of gene vectors keyed by transcription factor names.
\describe{
  \item{JUN}{integer vector of Entrez Gene IDs for genes targeted by JUN}
  ...
}}
\source{
\url{http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeRegTfbsClustered/}
}
\usage{
ENCODE
}
\description{
Human transcription factor targets estimated by selecting a subset of
ChIP-seq peaks with a score of 1000.
}
\details{
The ChIP-seq data used here is taken from sample \code{ENCFF001UUQ}.
There are many other samples that you might want to use instead.
}
\examples{
length(ENCODE)
ENCODE[1]
\donttest{hist(sapply(ENCODE, length))}
}
\references{
ENCODE Project Consortium. An integrated encyclopedia of DNA
elements in the human genome. Nature 489, 57-74 (2012).
(\href{http://www.ncbi.nlm.nih.gov/pubmed/22955616}{PubMed})
}
\keyword{datasets}

