% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ITFP}
\alias{ITFP}
\title{Transcription factor targets from ITFP.}
\format{A list of gene vectors keyed by transcription factor names.
\describe{
  \item{JUN}{character vector of genes targeted by JUN}
  ...
}}
\source{
\url{itfp.biosino.org/itfp/}
}
\usage{
ITFP
}
\description{
Predicted human transcription factor targets scraped from ITFP.
}
\keyword{datasets}

