% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Marbach2016}
\alias{Marbach2016}
\title{Transcription factor targets from RegulatoryCircuits.org}
\format{A list of gene vectors keyed by transcription factor names.}
\source{
\url{http://www.regulatorycircuits.org/}
}
\usage{
Marbach2016
}
\description{
This data is taken from the \code{synoviocyte.txt} file. There are many
other cell types available that you might want to use instead.
}
\examples{
length(Marbach2016)
Marbach2016[1]
Marbach2016[["STAT3"]]
\donttest{hist(sapply(Marbach2016, length))}

}
\references{
Marbach, D., Lamparter, D., Quon, G., Kellis, M., Kutalik, Z.
& Bergmann, S. Tissue-specific regulatory circuits reveal variable modular
perturbations across complex diseases. Nat. Methods 13, 366-370 (2016).
(\href{http://www.ncbi.nlm.nih.gov/pubmed/26950747}{PubMed})
}
\keyword{datasets}

