% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Neph2012}
\alias{Neph2012}
\title{Transcription factor targets from RegulatoryNetworks.org}
\format{A list of lists keyed by cell type names. Each sublist is keyed by
transcription factor names and returns a gene vector.}
\source{
\url{http://www.regulatorynetworks.org/}
}
\usage{
Neph2012
}
\description{
Human transcription factor targets obtained from DNaseI footprinting and TF
recognition sequences. Targets only include transcription factors.
}
\examples{
length(Neph2012)
Neph2012[["fHeart-DS12531"]][["STAT3"]]
\donttest{hist(sapply(Neph2012[[1]], length))}

}
\references{
Neph, S., Stergachis, A. B., Reynolds, A., Sandstrom, R.,
Borenstein, E. & Stamatoyannopoulos, J. A. Circuitry and dynamics of human
transcription factor regulatory networks. Cell 150, 1274-1286 (2012).
(\href{http://www.ncbi.nlm.nih.gov/pubmed/22959076}{PubMed})
}
\keyword{datasets}

