% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulatory_circuits}
\alias{regulatory_circuits}
\title{Transcription factor targets from Marbach et al. 2016.}
\format{A list of gene vectors keyed by transcription factor names.
\describe{
  \item{JUN}{character vector of genes targeted by JUN}
  ...
}}
\source{
\url{http://www.regulatorycircuits.org/}
}
\usage{
regulatory_circuits
}
\description{
This data is taken from the \code{synoviocyte.txt} file. There are many
other cell types available that you might want to use instead.
}
\examples{
length(regulatory_circuits)
regulatory_circuits[1]
\donttest{hist(sapply(regulatory_circuits, length))}
}
\references{
Marbach, D., Lamparter, D., Quon, G., Kellis, M., Kutalik, Z.
& Bergmann, S. Tissue-specific regulatory circuits reveal variable modular
perturbations across complex diseases. Nat. Methods 13, 366-370 (2016).
(\href{http://www.ncbi.nlm.nih.gov/pubmed/26950747}{PubMed})
}
\keyword{datasets}

