% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRRUST}
\alias{TRRUST}
\title{Transcription factor targets from TRRUST.}
\format{A list of gene vectors keyed by transcription factor names.
\describe{
  \item{JUN}{integer vector of Entrez Gene IDs for genes targeted by JUN}
  ...
}}
\source{
\url{http://www.grnpedia.org/trrust/}
}
\usage{
TRRUST
}
\description{
Transcription factor targets from TRRUST.
}
\examples{
length(TRRUST)
TRRUST[1]
\donttest{hist(sapply(TRRUST, length))}
}
\references{
Han, H., Shim, H., Shin, D., Shim, J. E., Ko, Y., Shin, J.,
Kim, H., Cho, A., Kim, E., Lee, T., Kim, H., Kim, K., Yang, S., Bae, D.,
Yun, A., Kim, S., Kim, C. Y., Cho, H. J., Kang, B., Shin, S. & Lee, I.
TRRUST: a reference database of human transcriptional regulatory
interactions. Sci. Rep. 5, 11432 (2015).
(\href{http://www.ncbi.nlm.nih.gov/pubmed/26066708}{PubMed})
}
\keyword{datasets}

