% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ENCODE}
\alias{ENCODE}
\title{Transcription factor targets estimated from ENCODE data.}
\format{A list of gene vectors keyed by transcription factor names.
\describe{
  \item{JUN}{integer vector of Entrez Gene IDs for genes targeted by JUN}
  ...
}}
\source{
\url{http://hgdownload.cse.ucsc.edu/goldenpath/hg19/encodeDCC/wgEncodeRegTfbsClustered/}
}
\usage{
ENCODE
}
\description{
Human transcription factor targets estimated by selecting a subset of
ChIP-seq peaks with a score of 1000.
}
\keyword{datasets}

