% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ITFP}
\alias{ITFP}
\title{Transcription factor targets from ITFP}
\format{A list of gene vectors keyed by transcription factor names.}
\source{
\url{itfp.biosino.org/itfp/}
}
\usage{
ITFP
}
\description{
Predicted human transcription factor targets scraped from ITFP.
}
\examples{
length(ITFP)
ITFP[1]
ITFP[["STAT3"]]
\donttest{hist(sapply(ITFP, length))}

}
\references{
Zheng, G., Tu, K., Yang, Q., Xiong, Y., Wei, C., Xie, L., Zhu,
Y. & Li, Y. ITFP: an integrated platform of mammalian transcription
factors. Bioinformatics 24, 2416-2417 (2008).
(\href{http://www.ncbi.nlm.nih.gov/pubmed/18713790}{PubMed})
}
\keyword{datasets}

