% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{TRED}
\alias{TRED}
\title{Transcription factor targets from TRED.}
\format{A list of gene vectors keyed by transcription factor names.
\describe{
  \item{JUN}{integer vector of Entrez Gene IDs for genes targeted by JUN}
  ...
}}
\source{
\url{https://cb.utdallas.edu/cgi-bin/TRED/}
}
\usage{
TRED
}
\description{
Predicted and known human transcription factor targets scraped from TRED.
}
\keyword{datasets}

